requirements.txt
torch==2.2.0
torchvision==0.17.0
numpy==1.26.4
pandas==2.2.2
matplotlib==3.8.4
seaborn==0.13.2
scikit-learn==1.4.1
Pillow==10.4.0
